Room2  Resolved
Added the missing keyword 'async' to the array, noted the duplicate request for one constant to be called instead of both.

Noticed that the function did not have the correct symbol in place to show the most recent book as required but it was showing the oldest book in the JSON file it was calling from.

Staying with the functions, the intersection function was incomplete as it was only showing the information from setA and not comparing setA and setB. added the .filter to show the common concept as required.